requirements.txt
Flask==0.10.1

# Application server for both development and production.
gunicorn==19.4.5

requests >=2.0

pandas >=0.23.0

matplotlib >=3.0.2

python-dateutil ==2.7.5

tweepy==3.7.0